corregidos los tipos en entities, hay que revisar la carga de venta a db

+++ src-tauri/entity/src/config.rs
@@ -6,8 +6,8 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "config")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub cantidad_productos: i32,
+    pub id: i64,
+    pub cantidad_productos: u8,
     pub formato_producto: String,
     pub modo_mayus: String,
     #[sea_orm(column_type = "Double")]
